[WIP] use winutil service functions in controlsvc for agent
